refactor: api call functions are now grouped into the api folder
