+++ src/api/config.js
@@ -1,0 +1,4 @@
+const apiUrl = "http://localhost:3500";
+export const CONFIG_API = {
+  DB: `${apiUrl}/api`,
+};
