+++ src/components/Dashboard/SecondaryList.jsx
@@ -3,7 +3,6 @@ import ListItemButton from "@mui/material/ListItemButton";
 import ListItemIcon from "@mui/material/ListItemIcon";
 import ListItemText from "@mui/material/ListItemText";
 import ListSubheader from "@mui/material/ListSubheader";
-import LocationOnIcon from "@mui/icons-material/LocationOn";
 import ChatIcon from "@mui/icons-material/Chat";
 
 import { Link as RouterLink } from "react-router-dom";
@@ -20,12 +19,6 @@ function SecondaryList() {
         </ListItemIcon>
         <ListItemText primary="Chat" />
       </ListItemButton>
-      {/* <ListItemButton component={RouterLink} to="map">
-        <ListItemIcon>
-          <LocationOnIcon />
-        </ListItemIcon>
-        <ListItemText primary="Mapa" />
-      </ListItemButton> */}
     </React.Fragment>
   );
 }
